Made LoCoBot installation script modular for sim

+++ robots/LoCoBot/install/locobot_install_all.sh
@@ -25,6 +25,16 @@ if [ $(dpkg-query -W -f='${Status}' librealsense2 2>/dev/null | grep -c "ok inst
         sudo apt-get update && sudo apt-get -y upgrade && sudo apt-get -y dist-upgrade
 fi
 
+if [ -z "$1" ]; then
+	echo "Full Installation Option Chosen"
+	INSTALL_TYPE="full"
+elif [ "$1" == "sim" ]; then
+	echo "Sim-only Installation Option Chosen"
+	INSTALL_TYPE="sim_only"
+else 
+	echo "Invalid commandline argument"
+	exit 1
+fi
 
 # STEP 1 - Install basic dependencies
 declare -a package_names=(
@@ -88,50 +98,51 @@ declare -a ros_package_names=(
 
 install_packages "${ros_package_names[@]}"
 
+if [ $INSTALL_TYPE == "full" ]; then
 
-# STEP 4 - Install camera (Intel Realsense D435)
-echo "Installing camera dependencies..."
+	# STEP 4 - Install camera (Intel Realsense D435)
+	echo "Installing camera dependencies..."
 
-# STEP 4A: Install librealsense
-if [ $(dpkg-query -W -f='${Status}' librealsense2 2>/dev/null | grep -c "ok installed") -eq 0 ]; then
-	sudo apt-key adv --keyserver keys.gnupg.net --recv-key C8B3A55A6F3EFCDE || sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-key C8B3A55A6F3EFCDE
-	sudo add-apt-repository "deb http://realsense-hw-public.s3.amazonaws.com/Debian/apt-repo xenial main" -u
-	sudo apt-get update
-	version="2.18.1-0~realsense0.568"
-	sudo apt-get -y install librealsense2-udev-rules=${version}
-	sudo apt-get -y install librealsense2-dkms=1.3.4-0ubuntu1
-	sudo apt-get -y install librealsense2=${version}
-	sudo apt-get -y install librealsense2-utils=${version}
-	sudo apt-get -y install librealsense2-dev=${version}
-	sudo apt-get -y install librealsense2-dbg=${version}
-fi
+	# STEP 4A: Install librealsense
+	if [ $(dpkg-query -W -f='${Status}' librealsense2 2>/dev/null | grep -c "ok installed") -eq 0 ]; then
+		sudo apt-key adv --keyserver keys.gnupg.net --recv-key C8B3A55A6F3EFCDE || sudo apt-key adv --keyserver hkp://keyserver.ubuntu.com:80 --recv-key C8B3A55A6F3EFCDE
+		sudo add-apt-repository "deb http://realsense-hw-public.s3.amazonaws.com/Debian/apt-repo xenial main" -u
+		sudo apt-get update
+		version="2.18.1-0~realsense0.568"
+		sudo apt-get -y install librealsense2-udev-rules=${version}
+		sudo apt-get -y install librealsense2-dkms=1.3.4-0ubuntu1
+		sudo apt-get -y install librealsense2=${version}
+		sudo apt-get -y install librealsense2-utils=${version}
+		sudo apt-get -y install librealsense2-dev=${version}
+		sudo apt-get -y install librealsense2-dbg=${version}
+	fi
 
-# STEP 4B: Install realsense2 SDK from source (in a separate catkin workspace)
-CAMERA_FOLDER=~/camera_ws
-if [ ! -d "$CAMERA_FOLDER/src" ]; then
-	mkdir -p $CAMERA_FOLDER/src
-	cd $CAMERA_FOLDER/src/
-	catkin_init_workspace
+	# STEP 4B: Install realsense2 SDK from source (in a separate catkin workspace)
+	CAMERA_FOLDER=~/camera_ws
+	if [ ! -d "$CAMERA_FOLDER/src" ]; then
+		mkdir -p $CAMERA_FOLDER/src
+		cd $CAMERA_FOLDER/src/
+		catkin_init_workspace
+	fi
+	if [ ! -d "$CAMERA_FOLDER/src/realsense" ]; then
+		cd $CAMERA_FOLDER/src/
+		git clone https://github.com/intel-ros/realsense.git
+		cd realsense
+		git checkout a036d81bcc6890658104a8de1cba24538effd6e3
+	fi
+	if [ -d "$CAMERA_FOLDER/devel" ]; then
+		rm -rf $CAMERA_FOLDER/devel
+	fi
+	if [ -d "$CAMERA_FOLDER/build" ]; then
+		rm -rf $CAMERA_FOLDER/build
+	fi
+	cd $CAMERA_FOLDER
+	catkin_make clean
+	catkin_make -DCATKIN_ENABLE_TESTING=False -DCMAKE_BUILD_TYPE=Release
+	catkin_make install
+	echo "source ~/camera_ws/devel/setup.bash" >> ~/.bashrc
+	source ~/camera_ws/devel/setup.bash
 fi
-if [ ! -d "$CAMERA_FOLDER/src/realsense" ]; then
-	cd $CAMERA_FOLDER/src/
-	git clone https://github.com/intel-ros/realsense.git
-	cd realsense
-	git checkout a036d81bcc6890658104a8de1cba24538effd6e3
-fi
-if [ -d "$CAMERA_FOLDER/devel" ]; then
-	rm -rf $CAMERA_FOLDER/devel
-fi
-if [ -d "$CAMERA_FOLDER/build" ]; then
-	rm -rf $CAMERA_FOLDER/build
-fi
-cd $CAMERA_FOLDER
-catkin_make clean
-catkin_make -DCATKIN_ENABLE_TESTING=False -DCMAKE_BUILD_TYPE=Release
-catkin_make install
-echo "source ~/camera_ws/devel/setup.bash" >> ~/.bashrc
-source ~/camera_ws/devel/setup.bash
-
 
 # STEP 5 - Setup catkin workspace
 echo "Setting up robot software..."
@@ -149,8 +160,10 @@ cd $LOCOBOT_FOLDER
 rosdep update 
 rosdep install --from-paths src -i -y
 cd $LOCOBOT_FOLDER/src/pyrobot/robots/LoCoBot/install
-chmod +x install_orb_slam2.sh
-source install_orb_slam2.sh
+if [ $INSTALL_TYPE == "full" ]; then
+	chmod +x install_orb_slam2.sh
+	source install_orb_slam2.sh
+fi
 cd $LOCOBOT_FOLDER
 if [ -d "$LOCOBOT_FOLDER/devel" ]; then
 	rm -rf $LOCOBOT_FOLDER/devel
@@ -163,15 +176,16 @@ echo "source $LOCOBOT_FOLDER/devel/setup.bash" >> ~/.bashrc
 source $LOCOBOT_FOLDER/devel/setup.bash
 
 
-# STEP 6 - Dependencies and config for calibration
-chmod +x src/pyrobot/robots/LoCoBot/locobot_navigation/orb_slam2_ros/scripts/gen_cfg.py
-rosrun orb_slam2_ros gen_cfg.py
-HIDDEN_FOLDER=~/.robot
-if [ ! -d "$HIDDEN_FOLDER" ]; then
-	mkdir ~/.robot
-	cp $LOCOBOT_FOLDER/src/pyrobot/robots/LoCoBot/locobot_calibration/config/default.json ~/.robot/
+if [ $INSTALL_TYPE == "full" ]; then
+	# STEP 6 - Dependencies and config for calibration
+	chmod +x src/pyrobot/robots/LoCoBot/locobot_navigation/orb_slam2_ros/scripts/gen_cfg.py
+	rosrun orb_slam2_ros gen_cfg.py
+	HIDDEN_FOLDER=~/.robot
+	if [ ! -d "$HIDDEN_FOLDER" ]; then
+		mkdir ~/.robot
+		cp $LOCOBOT_FOLDER/src/pyrobot/robots/LoCoBot/locobot_calibration/config/default.json ~/.robot/
+	fi
 fi
-
 
 # STEP 7 - Make a virtual env to install other dependencies (with pip)
 virtualenv_name="pyenv_pyrobot"
